Handle thumbnail galleries for screenshots and boxscans
Thumbnail gallery need to be constrained to a given size, and scroll to
the "active" thumbnail corresponding to the image being shown in the
carousel.

For the screenshots I had to resort to Javascript to handle the size, as
I could not find a way to constraint the height of the thumbnails column
to the same height of the main screenshot being displayed.