Change integration method between the legacy and Laravel site
Trying to get the legacy site running inside the Laravel one is too much
work, as the legacy site code expects to run at the root of the domain.
So instead we'll keep the legacy site under a sub-domain (e.g
`legacy.atarilegend.com`) and we'll link to it.

It will mean users will have to log-in twice, but that's a small price
to pay.