Fix infinite loop in `npm watch`
SASS was copied resources from `public/images/` onto themselves,
resulting in a new change being triggered, and so on.

To fix it move the images to the resources folder so that they are
copied by SASS only once. Also ignore the copied version in
`public/images/`.

Do the same for the fonts, but this time by grabbing them from NPM.