For #134: Search in menu software
Update search to also search on software. Add a tab in the search
results to separate game vs. software results.

Add software to the auto-complete with an icon to distinguish game vs.
software.

Add search bar in the menu index page.

Update auto-complete to directly navigate to the selected result.