Attempt to fix unit tests
When renaming columns with SQLite we have to add the new column then
copy the value. The newly added column is not nullable but no default
value is provided, so it's failing. Interestingly it's only failing in
GitHub actions and not on my computer...

As a workaround, make the column nullable when it's created.