Implement URL slugs for games
Instead of using the ID in the URL, use a slug. The slug is editable in
the new CPANEL. Add a canonical link to make sure Google and other
search engines only consider the slug URL. Add a redirect route so that
IDs will still work, but will redirect to the slug URL.

Aside: Also implement creation of new games in CPANEL + add button for
deletion of games (but not implemented yet as there are many entities to
cascade the deletion to). As part of this, add the timestamps columns to
the game table.

Aside: Fix the sort in the admin games autocomplete to be consistent with the
main site one.