Fixes #224: Disable chunking in filepond uploads
When using chunking it looks like the chunking never fully completes,
and the file is left in the temporary space with a temporary name, and
not the original upload filename, which in turns makes it impossible to
process as we don't have the extension.

Possibly a bug in the `laravel-filepond` dependency? For now, disable
chunking.